NormalMapping to MergeBranch

+++ 3DProject/3DProject/ShaderCreater.h
@@ -18,7 +18,7 @@ public:
 
 	GLuint getShaderProgramID()const;
 	void createShaders(std::string vertexShader, std::string geometryShader, std::string fragmentShader);
-	void setFloat(std::string name, int value) const;
+	void setFloat(std::string name, float value) const;
 };
